pyproject.toml
[tool.poetry]
name = "slurm-autoscale-tfe"
version = "0.5.1"
description = ""
authors = ["Félix-Antoine Fortin <felix-antoine.fortin@calculquebec.ca>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.6.2"
python-hostlist = "^1.21"
filelock = "^3.16.1"
requests = ">=2"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
slurm_suspend = "slurm_autoscale_tfe:suspend"
slurm_resume = "slurm_autoscale_tfe:resume"
slurm_resume_fail = "slurm_autoscale_tfe:resume_fail"
